requirements.txt
asgiref==3.2.10
atomicwrites==1.4.0
attrs==21.2.0
certifi==2021.10.8
charset-normalizer==2.0.7
colorama==0.4.4
Django==3.0.5
django-filter==21.1
djangorestframework==3.11.0
djangorestframework-simplejwt==5.0.0
gunicorn==20.1.0
idna==3.3
iniconfig==1.1.1
isort==5.9.3
packaging==21.0
pluggy==0.13.1
psycopg2-binary==2.8.6
py==1.10.0
PyJWT==2.3.0
pyparsing==2.4.7
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
python-dotenv==0.13.0
pytz==2020.1
requests==2.26.0
sqlparse==0.3.1
toml==0.10.2
urllib3==1.26.7